+++ glibc/math/__SVBool_t.c
@@ -1,0 +1,8 @@
+#include <math.h>
+
+int main(void)
+{
+	__SVBool_t b;
+	(void)b;
+	return 0;
+}