+++ glibc/math/__SVFloat64_t.c
@@ -1,0 +1,8 @@
+#include <math.h>
+
+int main(void)
+{
+	__SVFloat64_t f;
+	(void)f;
+	return 0;
+}